Port to JupyterLab v2
