+++ package.json
@@ -1,15 +1,15 @@
 {
   "name": "jupyterlab-favorites",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Add the ability to save favorite folders to JupyterLab for quicker browsing",
   "keywords": [
     "jupyter",
     "jupyterlab",
     "jupyterlab-extension"
   ],
-  "homepage": "https://github.com/tslaton/jupyterlab-favorites",
+  "homepage": "https://github.com/NERSC/jupyterlab-favorites",
   "bugs": {
-    "url": "https://github.com/tslaton/jupyterlab-favorites/issues"
+    "url": "https://github.com/NERSC/jupyterlab-favorites/issues"
   },
   "license": "BSD-3-Clause",
   "author": "Trevor Slaton",
@@ -21,27 +21,43 @@
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
   "style": "style/index.css",
+  "sideEffects": [
+    "style/**/*"
+  ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/tslaton/jupyterlab-favorites.git"
+    "url": "https://github.com/NERSC/jupyterlab-favorites.git"
   },
   "scripts": {
     "build": "tsc",
-    "clean": "rimraf lib && rimraf tsconfig.tsbuildinfo",
+    "clean": "rimraf lib tsconfig.tsbuildinfo",
     "prepare": "npm run clean && npm run build",
     "watch": "tsc -w"
   },
   "dependencies": {
-    "@jupyterlab/application": "^1.0.0",
-    "@jupyterlab/filebrowser": "^1.0.0",
-    "@jupyterlab/mainmenu": "^1.0.0"
+    "@jupyterlab/application": "^2.0.0",
+    "@jupyterlab/apputils": "^2.0.2",
+    "@jupyterlab/coreutils": "^4.0.2",
+    "@jupyterlab/docregistry": "^2.0.2",
+    "@jupyterlab/filebrowser": "^2.0.0",
+    "@jupyterlab/mainmenu": "^2.0.0",
+    "@jupyterlab/services": "^5.0.2",
+    "@jupyterlab/settingregistry": "^2.0.1",
+    "@jupyterlab/ui-components": "^2.0.2",
+    "@lumino/algorithm": "^1.2.3",
+    "@lumino/commands": "^1.10.1",
+    "@lumino/signaling": "^1.3.5",
+    "@lumino/widgets": "^1.11.1"
   },
   "devDependencies": {
     "rimraf": "^2.6.1",
-    "typescript": "~3.5.2"
+    "typescript": "~3.7.3"
   },
   "jupyterlab": {
     "extension": true,
     "schemaDir": "schema"
+  },
+  "prettier": {
+    "singleQuote": true
   }
 }
