+++ style/index.css
@@ -1,22 +1,3 @@
-:root {
-  /* All of the avaialble icons */
-  --jp-icon-favorite-filled-light: url('icons/md/light/baseline-star-24px.svg');
-  --jp-icon-favorite-unfilled-light: url('icons/md/light/baseline-star_border-24px.svg');
-  --jp-icon-favorite-filled-dark: url('icons/md/dark/baseline-star-24px.svg');
-  --jp-icon-favorite-unfilled-dark: url('icons/md/dark/baseline-star_border-24px.svg');
-  /* The icons chosen, based on current theme */
-  --jp-icon-favorite-filled: var(--jp-icon-favorite-filled-light);
-  --jp-icon-favorite-unfilled: var(--jp-icon-favorite-unfilled-light);
-}
-
-.jp-FavoritesIcon-filled,
-.jp-FavoritesIcon-unfilled {
-  min-width: 16px;
-  min-height: 16px;
-  background-size: 16px;
-  vertical-align: sub;
-}
-
 .jp-Favorites-pinner {
   position: absolute;
   right: 6px;
@@ -29,12 +10,14 @@
   margin: 4px 26px 4px 12px;
 }
 
-.jp-FavoritesIcon-filled {
-  background-image: var(--jp-icon-favorite-filled);
-}
-
-.jp-FavoritesIcon-unfilled {
-  background-image: var(--jp-icon-favorite-unfilled);
+.jp-Favorites-BreadCrumbs-Icon svg {
+  border-radius: var(--jp-border-radius);
+  cursor: pointer;
+  height: 16px;
+  margin: 0px 2px;
+  padding: 0px 2px;
+  vertical-align: middle;
+  width: 16px;
 }
 
 .jp-Favorites-header,
@@ -69,9 +52,11 @@
   background-color: var(--jp-layout-color2);
 }
 
-.jp-Favorites-itemIcon {
-  flex: 0 0 20px;
-  margin-right: 4px;
+.jp-Favorites-itemIcon svg {
+  display: block;
+  height: 16px;
+  margin: 0 5px;
+  width: 16px;
 }
 
 .jp-Favorites-itemText {
